disease prediction/api/requirements.txt
pillow
flask==2.0.3
tensorflow==2.6.2